Update resolution.h
High resolution for position, velocity and torque

+++ vulp/actuation/resolution.h
@@ -32,9 +32,9 @@ inline actuation::moteus::QueryCommand get_query_resolution() {
   using actuation::moteus::Resolution;
   actuation::moteus::QueryCommand query;
   query.mode = Resolution::kInt16;
-  query.position = Resolution::kInt16;
-  query.velocity = Resolution::kInt16;
-  query.torque = Resolution::kInt16;
+  query.position = Resolution::kInt32;
+  query.velocity = Resolution::kInt32;
+  query.torque = Resolution::kInt32;
   query.q_current = Resolution::kInt16;
   query.d_current = Resolution::kInt16;
   query.rezero_state = Resolution::kIgnore;
